+++ database/models/__init__.py
@@ -1,2 +1,3 @@
 from .tyre_model import TyreModel
 from .tyre_impression import TyreImpression
+from .tyre_impression_processing import TyreImpressionProcessing