+++ database/repositories/__init__.py
@@ -1,2 +1,3 @@
 from .tyre_model_repository import TyreModelRepository
 from .tyre_impression_repository import TyreImpressionRepository
+from .tyre_impression_processing_repository import TyreImpressionProcessingRepository