+++ database/repositories/tyre_impression_processing_repository.py
@@ -1,0 +1,10 @@
+from database.models import TyreImpressionProcessing
+from database.repositories.base_repository import BaseRepository
+
+
+class TyreImpressionProcessingRepository(BaseRepository[TyreImpressionProcessing]):
+
+    def __init__(self):
+        super().__init__(TyreImpressionProcessing)
+
+